抽取Spring API为公共抽象父类
