+++ powerjob-worker/src/main/java/tech/powerjob/worker/processor/impl/BuiltInSpringProcessorFactory.java
@@ -19,17 +19,11 @@ import java.util.Set;
  * @since 2023/1/17
  */
 @Slf4j
-public class BuiltInSpringProcessorFactory implements ProcessorFactory {
+public class BuiltInSpringProcessorFactory extends AbstractBuildInSpringProcessorFactory {
 
-    private final ApplicationContext applicationContext;
 
     public BuiltInSpringProcessorFactory(ApplicationContext applicationContext) {
-        this.applicationContext = applicationContext;
-    }
-
-    @Override
-    public Set<String> supportTypes() {
-        return Sets.newHashSet(ProcessorType.BUILT_IN.name());
+        super(applicationContext);
     }
 
     @Override
@@ -59,42 +53,5 @@ public class BuiltInSpringProcessorFactory implements ProcessorFactory {
         return null;
     }
 
-    private boolean checkCanLoad() {
-        try {
-            ApplicationContext.class.getClassLoader();
-            return applicationContext != null;
-        } catch (Throwable ignore) {
-        }
-        return false;
-    }
-
-
-    @SuppressWarnings("unchecked")
-    private static <T> T getBean(String className, ApplicationContext ctx) throws Exception {
-
-        // 0. 尝试直接用 Bean 名称加载
-        try {
-            final Object bean = ctx.getBean(className);
-            if (bean != null) {
-                return (T) bean;
-            }
-        } catch (Exception ignore) {
-        }
-
-        // 1. ClassLoader 存在，则直接使用 clz 加载
-        ClassLoader classLoader = ctx.getClassLoader();
-        if (classLoader != null) {
-            return (T) ctx.getBean(classLoader.loadClass(className));
-        }
-        // 2. ClassLoader 不存在(系统类加载器不可见)，尝试用类名称小写加载
-        String[] split = className.split("\\.");
-        String beanName = split[split.length - 1];
-        // 小写转大写
-        char[] cs = beanName.toCharArray();
-        cs[0] += 32;
-        String beanName0 = String.valueOf(cs);
-        log.warn("[SpringUtils] can't get ClassLoader from context[{}], try to load by beanName:{}", ctx, beanName0);
-        return (T) ctx.getBean(beanName0);
-    }
 
 }
